Added drop if exists command before creating view.

+++ database/CS6232-G1-create_DB.sql
@@ -396,7 +396,8 @@ ALTER DATABASE [CS6232-G1] SET  READ_WRITE
 GO
 
 GO
-DROP VIEW dbo.EvaluationScheduleView
+IF OBJECT_ID('dbo.EvaluationScheduleView', 'V') IS NOT NULL
+    DROP VIEW dbo.EvaluationScheduleView;
 GO
 CREATE VIEW dbo.EvaluationScheduleView
 AS
